ai_tools_monitor/requirements.txt
# Core dependencies
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
playwright==1.42.0
aiohttp==3.9.3
sqlalchemy==2.0.28
pydantic==2.6.3

# LLM and AI
openai==1.12.0
langchain==0.1.12
langchain-openai==0.0.8
crewai==0.28.0

# Data processing
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
nltk==3.8.1

# Web scraping and automation
selenium==4.18.1
tweepy==4.14.0
praw==7.7.1
producthunt==0.0.1

# Database
psycopg2-binary==2.9.9
pinecone-client==3.0.2

# API and web
fastapi==0.110.0
uvicorn==0.27.1
streamlit==1.32.0
python-multipart==0.0.9

# Monitoring and scheduling
schedule==1.2.1
python-jose==3.3.0
passlib==1.7.4

# Testing and development
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
black==24.2.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Web Scraping
aiohttp>=3.8.0
beautifulsoup4>=4.9.3
playwright>=1.40.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# API Clients
praw>=7.7.0  # Reddit API
tweepy>=4.15.0  # Twitter API
requests>=2.31.0  # HTTP client
beautifulsoup4>=4.12.0  # HTML parsing
playwright>=1.42.0  # Browser automation
aiohttp>=3.9.3  # Async HTTP client

# Database
sqlalchemy>=2.0.28  # ORM
pydantic>=2.6.3  # Data validation

# AI/ML
openai>=1.12.0  # OpenAI API
langchain>=0.1.12  # LangChain
langchain-openai>=0.0.8  # LangChain OpenAI integration

# UI
streamlit>=1.32.0  # Web interface

# Utilities
python-dotenv>=1.0.0  # Environment variables
asyncio>=3.4.3
aiofiles>=23.2.1
tenacity>=8.2.0  # For retries
fake-useragent>=1.4.0  # For rotating user agents

# Database (if needed later)
# sqlalchemy>=2.0.0
# alembic>=1.12.0

# MCP Claude Code and its dependencies
mcp-claude-code>=0.3.4
litellm>=1.0.0
jupyter>=1.0.0 